moved gtm tag to document.tsx per task 6150
